requirements.txt
Flask==3.0.3
fuzzywuzzy==0.18.0
mysql-connector-python==9.2.0
nltk==3.8.1
numpy==2.2.0
pandas==2.2.2
python-dotenv==1.0.1
rouge_score==0.1.2
scikit-learn==1.5.1
sentence-transformers==3.0.1
huggingface-hub==0.23.5
haystack-ai==2.7.0
chroma-haystack==2.0.0
chromadb==0.6.3
json5==0.9.25

